+++ CHANGELOG.md
@@ -9,5 +9,6 @@ To get the diff between two versions, go to https://github.com/xuanquynh/php-cod
 
 ### [Unreleased] v1.0.0
 
-  * feature Add SunAsterisk Standards.
-  * feature Add XuanQuynh Standards.
+  * feature Initialize Laravel Standards.
+  * feature Initialize SunAsterisk Standards.
+  * feature Initialize XuanQuynh Standards.